pcocam2: Fixed incorrect parameter size.

+++ pcocam2App/src/SimulationApi.cpp
@@ -1241,8 +1241,8 @@ int SimulationApi::doSetNoiseFilterMode(Handle handle, unsigned short mode)
 /**
  * Set the camera RAM segment size
  */
-int SimulationApi::doSetCameraRamSegmentSize(Handle handle, unsigned short seg1,
-	unsigned short seg2, unsigned short seg3, unsigned short seg4)
+int SimulationApi::doSetCameraRamSegmentSize(Handle handle, unsigned long seg1,
+	unsigned long seg2, unsigned long seg3, unsigned long seg4)
 {
     return DllApi::errorNone;
 }
